requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
frozenlist==1.3.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
inflection==0.5.1
jsonschema==4.16.0
lxml==4.9.1
model-bakery==1.9.0
multidict==6.0.2
psycopg2==2.9.5
pyrsistent==0.18.1
pytz==2022.4
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
sentry-sdk==1.10.1
sniffio==1.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.8.1
